float values wrongly set to long

+++ EnvironmentMonitor/Bme280Wrapper.cpp
@@ -28,25 +28,25 @@ void Bme280Wrapper::init() {
     }
 }
 
-long Bme280Wrapper::readTempC() {
+float Bme280Wrapper::readTempC() {
     return myBME280.readTempC();
 }
 
-long Bme280Wrapper::readTempF() {
+float Bme280Wrapper::readTempF() {
     return myBME280.readTempF();
 }
 
-long Bme280Wrapper::readFloatHumidity() {
+float Bme280Wrapper::readFloatHumidity() {
     return myBME280.readFloatHumidity();
 }
 
-long Bme280Wrapper::readFloatPressure() {
+float Bme280Wrapper::readFloatPressure() {
     return myBME280.readFloatPressure();
 }
 
-long Bme280Wrapper::readFloatAltitudeMeters() {
+float Bme280Wrapper::readFloatAltitudeMeters() {
     return myBME280.readFloatAltitudeMeters();
 }
-long Bme280Wrapper::readFloatAltitudeFeet() {
+float Bme280Wrapper::readFloatAltitudeFeet() {
     return myBME280.readFloatAltitudeFeet();
 }
